Improve events handling, add in global event styles (GWH non-location stuff)
