+++ src/modules/location-hud/region-events/great-winter-hunt/index.js
@@ -1,0 +1,7 @@
+import { addHudStyles } from '@/utils';
+
+import styles from './styles.css';
+
+export default () => {
+  addHudStyles(styles);
+};
