+++ src/modules/location-hud/region-events/halloween/index.js
@@ -1,0 +1,25 @@
+import { addHudStyles, onRequest } from '@/utils';
+
+import styles from './styles.css';
+
+const undisableCheese = () => {
+  const armButtons = document.querySelectorAll('.halloweenBoilingCauldronHUD-bait');
+  armButtons.forEach((armButton) => {
+    armButton.classList.remove('disabled');
+    const link = armButton.querySelector('a.disabled');
+    if (link) {
+      link.classList.remove('disabled');
+    }
+
+    const tooltipLink = armButton.querySelector('.halloweenBoilingCauldronHUD-bait-tooltipContent a.disabled');
+    if (tooltipLink) {
+      tooltipLink.classList.remove('disabled');
+    }
+  });
+};
+
+export default () => {
+  addHudStyles(styles);
+  undisableCheese();
+  onRequest(undisableCheese);
+};
